+++ odesk/__init__.py
@@ -13,7 +13,7 @@ so you can use::
 
 """
 
-VERSION = '0.5.4'
+VERSION = '0.5.5'
 
 
 def get_version():